Consistently use long-typed milliseconds everywhere
Testing: All unit tests pass.

Affects: https://jira.nypl.org/browse/SIMPLY-1302

+++ org.nypl.audiobook.android.open_access/src/main/java/org/nypl/audiobook/android/open_access/ExoAudioBookPlayer.kt
@@ -792,10 +792,10 @@ class ExoAudioBookPlayer private constructor(
      */
 
     if (requestedSpineElement == currentSpineElement) {
-      this.seek(location.offsetMilliseconds.toLong())
+      this.seek(location.offsetMilliseconds)
       this.opPlay()
     } else {
-      this.playSpineElementIfAvailable(requestedSpineElement, location.offsetMilliseconds.toLong())
+      this.playSpineElementIfAvailable(requestedSpineElement, location.offsetMilliseconds)
     }
   }
